cp-style multi-create
This doesn't support recursive operations yet, but I often want to
move a few files in a directory to a remote object container, so this
will do that in a sensible way.

I couldn't figure out how to fail optparse-applicative parsers after
parsing a bunch of stuff with a some or many, so I just parse it into
an Either String and then report the error later.  That's not great,
but it's not a terrible experience.